feat: v2.6.0 implemented profile handle system

+++ components/FollowButton.tsx
@@ -4,12 +4,13 @@ import useAuth from "@/hooks/useAuth";
 import { useDict } from "@/hooks/useDict";
 import { useToast } from "@/hooks/useToast";
 import { supabase } from "@/lib/supabase";
-import { useQuery, useQueryClient } from "@tanstack/react-query";
+import { skipToken, useQuery, useQueryClient } from "@tanstack/react-query";
 import { useRouter } from "next/navigation";
 import React, { useState } from "react";
 import checkIsFollowing from "../queries/checkIsFollowing";
+import useProfile from "@/hooks/useProfile";
 
-export default function FollowButton({ userId, style }: { userId: string; style?: React.CSSProperties; }) {
+export default function FollowButton({ userId, style }: { userId: string | undefined; style?: React.CSSProperties; }) {
     const { user } = useAuth();
     const { toast } = useToast();
     const router = useRouter();
@@ -19,10 +20,11 @@ export default function FollowButton({ userId, style }: { userId: string; style?
 
     const [hoverFollowButton, setHoverFollowButton] = useState(false);
 
+    const { data: profile, isLoading, error } = useProfile(userId);
+
     const { data: isFollowing, isLoading: isFollowingLoading } = useQuery({
-        queryKey: ["isFollowing", user?.id, userId],
-        queryFn: () => checkIsFollowing(user?.id, userId),
-        enabled: !!user?.id && !!userId,
+        queryKey: ["isFollowing", user?.id, profile?.id],
+        queryFn: profile?.id ? () => checkIsFollowing(user?.id, profile?.id) : skipToken,
         staleTime: Infinity
     });
 
@@ -31,34 +33,42 @@ export default function FollowButton({ userId, style }: { userId: string; style?
             router.push(`/${dict.langSubTag}/main/signin`);
             return;
         }
+        if (!profile?.id) {
+            toast(undefined, dict.error.getUrl, 'errorDb');
+            return;
+        }
         const { error } = await supabase
             .from('follows')
-            .insert({ follower_id: user?.id, following_id: userId });
+            .insert({ follower_id: user.id, following_id: profile.id });
         if (error) {
             console.error("Failed to follow a user:", error);
             toast(undefined, dict.error.updateProfile, 'errorDb');
             return;
         }
-        queryClient.setQueryData(['isFollowing', user?.id, userId], true);
-        queryClient.invalidateQueries({ queryKey: ['profileCount', userId] });
+        queryClient.setQueryData(['isFollowing', user.id, profile.id], true);
+        queryClient.invalidateQueries({ queryKey: ['profileCount', profile.id] });
     }
 
     async function handleUnfollow() {
+        if (!profile?.id) {
+            toast(undefined, dict.error.getUrl, 'errorDb');
+            return;
+        }
         const { error } = await supabase
             .from('follows')
             .delete()
             .eq('follower_id', user?.id)
-            .eq('following_id', userId);
+            .eq('following_id', profile.id);
         if (error) {
             console.error("Failed to unfollow a user:", error);
             toast(undefined, dict.error.updateProfile, 'errorDb');
             return;
         }
-        queryClient.setQueryData(['isFollowing', user?.id, userId], false);
-        queryClient.invalidateQueries({ queryKey: ['profileCount', userId] });
+        queryClient.setQueryData(['isFollowing', user?.id, profile.id], false);
+        queryClient.invalidateQueries({ queryKey: ['profileCount', profile.id] });
     }
 
     return (
-        <button type="button" style={style} onClick={isFollowing ? handleUnfollow : handleFollow} onMouseEnter={() => setHoverFollowButton(true)} onMouseLeave={() => setHoverFollowButton(false)} className={`${isFollowing ? hoverFollowButton ? "text-rose-500 bg-rose-400/20 border border-rose-400" : 'bg-foreground border border-muted opacity-60' : 'bg-foreground'} px-6 py-2 text-xs rounded-full font-bold text-accent max-w-sm inline-block transition-all duration-100 ease-out`}>{isFollowingLoading ? dict.ui.loading : isFollowing ? hoverFollowButton ? dict.ui.unfollow : dict.profile.following : dict.ui.follow}</button>
+        <button type="button" style={style} onClick={isFollowing ? handleUnfollow : handleFollow} onMouseEnter={() => setHoverFollowButton(true)} onMouseLeave={() => setHoverFollowButton(false)} className={`${isFollowing ? hoverFollowButton ? "text-rose-500 bg-rose-400/20 border border-rose-400" : 'bg-foreground border border-muted opacity-60' : 'bg-foreground'} px-6 py-2 text-xs rounded-full font-bold text-accent max-w-sm inline-block transition-all duration-100 ease-out`}>{isFollowingLoading || isLoading ? dict.ui.loading : isFollowing ? hoverFollowButton ? dict.ui.unfollow : dict.profile.following : dict.ui.follow}</button>
     );
 }
